Changes + README.md
