+++ literki/literki.py
@@ -1,40 +1,3 @@
-# Importy bibliotek:
-# Depending on the OS - Windows or Linux - we pick getch (linux) or msvcrt.
-from pygame import mixer
-from getch import getch
-import time
-import os
-# import keyboard
-from glob import glob
-
-# Audio = dzwiek
-
-# Ustawienia dzwieku:
-mixer.init()
-
-# Funkcja do grania dzwieków:
-def graj(nazwa_pliku):
-    filename_with_path = f"audio/{nazwa_pliku}.mp3";
-    if ( glob(filename_with_path) == [] ):
-        print (f"Blad: Nie znalazlem pliku: {nazwa_pliku}. Nie gram nic.")
-        return
-    print (f"Gram {nazwa_pliku}")
-
-    mixer.music.load(filename_with_path)
-    mixer.music.play()
-    while mixer.music.get_busy():  # wait for music to finish playing
-        time.sleep(1)
-
-# Wypisuje wszystkie pliki:
-def wypisz_pliki_audio():
-    pliki = glob("audio/*.mp3")
-    pliki.sort()
-    for plik in pliki:
-        p = plik.replace("audio/", "")
-        p = p.replace(".mp3", "")
-        print(p)
-
-#
 # Twoj kod:
 # 
 # Funkcje:
@@ -66,5 +29,15 @@ def wypisz_pliki_audio():
 
 #graj("a")
 
-klawisz = getch()
-print(f"Wcisnal(les/las) {klawisz}")
+from funkcje import *
+
+# while(True):
+#     klawisz = getch().lower()
+#     print(f"Wcisnal(les/las) {klawisz}")
+#     graj(klawisz.decode())
+
+
+for i in range(0, 10):
+    print(i)
+
+graj('dobrze7')
